otlp code dups
