+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_http_log_exporter.h
@@ -9,6 +9,7 @@
 #  include "opentelemetry/exporters/otlp/otlp_http_client.h"
 
 #  include "opentelemetry/exporters/otlp/otlp_environment.h"
+#  include "opentelemetry/exporters/otlp/otlp_http_exporter_options.h"
 
 #  include <chrono>
 #  include <memory>
@@ -19,38 +20,6 @@ namespace exporter
 {
 namespace otlp
 {
-
-/**
- * Struct to hold OTLP exporter options.
- */
-struct OtlpHttpLogExporterOptions
-{
-  // The endpoint to export to. By default
-  // @see
-  // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/otlp.md
-  // @see https://github.com/open-telemetry/opentelemetry-collector/tree/main/receiver/otlpreceiver
-  std::string url = GetOtlpDefaultHttpLogEndpoint();
-
-  // By default, post json data
-  HttpRequestContentType content_type = HttpRequestContentType::kJson;
-
-  // If convert bytes into hex. By default, we will convert all bytes but id into base64
-  // This option is ignored if content_type is not kJson
-  JsonBytesMappingKind json_bytes_mapping = JsonBytesMappingKind::kHexId;
-
-  // If using the json name of protobuf field to set the key of json. By default, we will use the
-  // field name just like proto files.
-  bool use_json_name = false;
-
-  // Whether to print the status of the exporter in the console
-  bool console_debug = false;
-
-  // TODO: Enable/disable to verify SSL certificate
-  std::chrono::system_clock::duration timeout = GetOtlpDefaultLogTimeout();
-
-  // Additional HTTP headers
-  OtlpHeaders http_headers = GetOtlpDefaultLogHeaders();
-};
 
 /**
  * The OTLP exporter exports log data in OpenTelemetry Protocol (OTLP) format.
@@ -67,7 +36,7 @@ public:
    * Create an OtlpHttpLogExporter with user specified options.
    * @param options An object containing the user's configuration options.
    */
-  OtlpHttpLogExporter(const OtlpHttpLogExporterOptions &options);
+  OtlpHttpLogExporter(const OtlpHttpExporterOptions &options);
 
   /**
    * Creates a recordable that stores the data in a JSON object
@@ -94,7 +63,7 @@ private:
   friend class OtlpHttpLogExporterTestPeer;
 
   // Configuration options for the exporter
-  const OtlpHttpLogExporterOptions options_;
+  const OtlpHttpExporterOptions options_;
 
   // Object that stores the HTTP sessions that have been created
   OtlpHttpClient http_client_;
